feat(wallet-sdk): own the encryption core (connection-core step 2)
Move the framework-free encryption layer out of features/shared/encryption.ts into
@agicash/wallet-sdk — the first connection-core piece the SDK owns on the path to backing the
accounts domain with SDK-owned connections.

- packages/wallet-sdk/src/encryption.ts: ECIES wrappers (encryptToPublicKey/
  decryptWithPrivateKey + batch variants), the serialization rules (Date/undefined/non-finite/
  Money preservation), the Encryption type, getEncryption, and the opensecret-backed key
  derivation queryOptions (encryptionPrivateKey/PublicKey) — moved verbatim except:
  - imports repointed (~/lib/ecies -> @agicash/utils/ecies, ~/lib/money -> @agicash/utils/money)
  - the two queryOptions are now PLAIN query-options objects (the react-query queryOptions()
    helper is a typed identity fn — runtime-identical; web's useSuspenseQuery consumes them
    unchanged, per the SDK's query-core-types-only architecture)
- web's features/shared/encryption.ts keeps ONLY the React hooks (useEncryption,
  useEncryptionPrivateKey/PublicKeyHex), now consuming the SDK; everything else re-exported
  (all 12 consumer files untouched)
- SDK gains @agicash/opensecret (catalog) — the single TRACKED EXCEPTION to the framework-free
  rule (transitive react peer-dep + localStorage), documented in package.json
  dependencies:comments; removed by the in-flight storage-pluggable opensecret bump before the
  MCP phase. No DIRECT react/DOM imports in SDK src (gate verified).
- SDK barrel now documents the queryOptions-read pattern + the deferred Query<T>/MCP plan
- new encryption.test.ts locks the encrypt/decrypt roundtrip + the serialization rules

typecheck, biome, build all green; 138 tests (sdk 5, web 84, utils 32, cashu 35).
